Grab location from device and move map.
Activities are renamed so that this fragment can be added to a future
"Home" activity.

TODO: Write location to disk to simulate sending to a server.
TODO: Build an actual "Home" activity for user interaction.
