+++ app/src/main/res/layout/activity_map.xml
@@ -6,4 +6,9 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home" />
+    map:uiRotateGestures="false"
+    map:uiScrollGestures="true"
+    map:uiTiltGestures="false"
+    map:uiZoomControls="true"
+    map:uiZoomGestures="true"
+    tools:context=".Map" />
